生成员工授信结算单，完成结算

+++ persist/src/main/java/com/welfare/persist/entity/EmployeeSettleDetail.java
@@ -58,7 +58,7 @@ public class EmployeeSettleDetail extends Model<EmployeeSettleDetail> implements
      * 账户
      */
     @ApiModelProperty("账户")   
-    private Integer accountCode;
+    private Long accountCode;
     /**
      * 账户名称
      */
